basic handshake works with my open router!
I tested this on an open router and it worked. I have no idea if this will work with *every* router, since I did not follow the spec fully.

Right now, I only send the fields which the router complained about. There are many fields which I do not send that are technically required by the spec.